+++ packages/vendure-plugin-metrics/src/api/cache.ts
@@ -1,16 +1,32 @@
+interface EntryDate {
+  createdAt: Date;
+}
+
 /**
  * Very basic in memory cache for storing metric results
  */
 export class Cache<T> {
-  private cache = new Map<string, T>();
-  private key(key: any): string {
-    return JSON.stringify(key);
-  }
+  // Default max age is 12 hours
+  constructor(private maxAgeInSeconds: number = 60 * 60 * 12) {}
 
-  set(key: any, value: T): void {
-    this.cache.set(this.key(key), value);
+  private cache = new Map<string, T & EntryDate>();
+
+  set(key: string, value: T): void {
+    this.cache.set(key, {
+      ...value,
+      createdAt: new Date(),
+    });
   }
-  get(key: any): T | undefined {
-    return this.cache.get(this.key(key));
+  get(key: string): T | undefined {
+    const res = this.cache.get(key);
+    if (!res) {
+      return undefined;
+    }
+    const now = new Date();
+    if (now.getTime() - res.createdAt.getTime() > this.maxAgeInSeconds * 1000) {
+      this.cache.delete(key);
+      return undefined;
+    }
+    return res;
   }
 }