+++ packages/vendure-plugin-metrics/src/index.ts
@@ -3,4 +3,5 @@ export * from './api/metrics.service';
 export * from './api/metrics.resolver';
 export * from './ui/generated/graphql';
 export * from './api/metrics/average-order-value';
-export * from './api/metrics/sales-per-product';
+export * from './api/metrics/revenue-per-product';
+export * from './api/metrics/units-sold-metric';